+++ src/vtex/getProductSimilars/index.ts
@@ -1,0 +1,3 @@
+import { getProductSimilars } from './getProductSimilars';
+
+export default getProductSimilars;